Folded tests into correct larger tests and added new tests to SGPWorld.test.cc
