+++ source/catch/main.cc
@@ -34,10 +34,7 @@
 #include "../test/sgp_mode_test/SGPHost.test.cc"
 #include "../test/sgp_mode_test/SGPSymbiont.test.cc"
 #include "../test/sgp_mode_test/StressHost.test.cc"
-#include "../test/sgp_mode_test/FirstCredit.test.cc"
-#include "../test/sgp_mode_test/OnlyNOT.test.cc"
-#include "../test/sgp_mode_test/ParentTask.test.cc"
-//#include "../test/sgp_mode_test/HealthHost.test.cc"
+#include "../test/sgp_mode_test/HealthHost.test.cc"
 
 #include "../test/integration_test/spatial_structure/vt.test.cc"
 #include "../test/integration_test/lysogeny/plr.test.cc"
